tests: use tap output and prove to run tests
trap EXIT in tests, and use that to output `ok` or `not ok` (based on
exit code), along with a tap test plan

replace manual test runner with a file in `scripts/` and use that in ci

return as ok when skipping, but with tap directive. output directives in
`prove`. skip `sudo` test if not root (instead of failing)